Move applicationState to a property instead
This simplifies the code in CrashLogging a bit so it doesn't need to muck
around with `#if os()`.
